+++ scenario-compiler/src/test/scala/pl/touk/nussknacker/engine/spel/SpelExpressionSpec.scala
@@ -1504,11 +1504,13 @@ class SpelExpressionSpec extends AnyFunSuite with Matchers with ValidatedValuesD
 
   test("should allow using list methods on array projection") {
     evaluate[Any]("'a,b'.split(',').![#this].isEmpty()") shouldBe false
+    evaluate[String]("'a,b'.split(',').![#this].get(0)") shouldBe "a"
   }
 
   test("should allow using list methods on array") {
     evaluate[Any]("#array.isEmpty()") shouldBe false
     evaluate[Any]("#intArray.isEmpty()") shouldBe false
+    evaluate[String]("#array.get(0)") shouldBe "a"
   }
 
   test("should allow using list methods on nested arrays") {